feat: added custom signer functionality
(1) In the `.signer()` method on the `.actionBuilder()` and `.dbBuilder()`, you can now pass a
callback function that returns a Uint8Array and the enumerator for the signature type. This allows
developers to use signing functions beyond just EtherJS.

(2) The `.nearConfig()` method was removed.
Because developers can now pass their own signing functions, there is no need for the config, as
developers can just pass the signing function from the `near-api-js` library.

BREAKING CHANGE: The `.nearConfig()` method has been removed from the `.actionBuilder()` and
`.dbBuilder()` classes, in favor of allowing developers to pass their own signing function to the
'.signer()` methods.