Feature/custom signatures (#31)
* feat: friendly signatures for transactions

Friendly signatures implemented on transactions. Developers can now customize the first part of the
signature message by chaining a `.description()` method to the `.actionBuilder()` or `.dbBuilder()`
when building a tx.

re #24

* feat: friendly signatures for transactions

Friendly signatures implemented on transactions. Developers can now customize the first part of the
signature message by chaining a `.description()` method to the `.actionBuilder()` or `.dbBuilder()`
when building a tx.

re #24

* fix: fix all xToBytes functions to take any ASCII character and up to Uint64 numbers

* refactor: cleanup leftover content from merge

cleaned up a few repeat methods and minor styling issues that were included in previous commit's
merge

* feat: add friendly signatures to actions sent to the `.call()` api

Similar to the `.description()` feature that was added for transactions in commit 0c625cd,
developers can now pass custom signature messages for actions that require signatures (`view`
actions that are built with `.buildMsg()`).

re #24

* docs: minor fixes on reamde and migration