feat(client/kwil.ts): require chainID to be configured in Kwil constructor
In the `WebKwil` or `NodeKwil` constructor, you must now pass a `chainId` property for the Kwil
chain you are connecting to. The chainId is included in the transaction payload when state-changing
transactions are built / executed.

BREAKING CHANGE: The `WebKwil` and `NodeKwil` classes require an additional `chainId` config string
to execute database deploys, database drops, or state-changing actions. You can check the chainId
for your kwilProvider by calling `kwil.chainInfo()`.

re #39