feat: make kwil.call() auto authenticate, when required
When using Kwil Gateway (kgw) for authentication, the `.call()` method will automatically try to
authenticate if it receives a 401 error. This removes the need for the `.authenticate()` and
`.setCookie()` methods.

BREAKING CHANGE: Remove the `kwil.authenticate()` and `kwil.setCookie()` methods, as those methods
are now handled internally by the `kwil.call()` method.