fix(rlp): fix bug where boolean false was not rlp encoded correctly
This commit fixes two bugs: (1) when passing a schema with public set to false, it was overriding
the property to true in the DbBuilder's `makePayloadEncodable` method. Rather than checking if the
property is falsy, it should just check if it is null or undefined. (2) Previously, `false` was
hexlified to `0x00` before RLP encoding, but the proper false hex encoding for RLP is `0x`.