fix: rlp encoding zero and case sensitivity for action input names
This commit fixes two bugs in kwil-js. First, handles how we RLP encode zero. Because kwil-db stores
all values as strings under the hood, we need to be RLP encoding the integer zero as a string. The
reason we only hit this bug on the number zero is that previously, the SDK coverted all truthy
values to a string, which missed converting zero to string, thereby sending zero to be rlp encoded
as an integer. This commit changes the stringify logic for all values that return with .toString(),
so zero does not get missed. Second, actionInput keys were not case insensitive, which led some
users to get unexpected errors when typing an action name as case sensitive. This commit forces all
actioninput keys to be lowercase.