feat: add procedures to schema
Made necessary adjustments to the `Database` and `CompiledKuneiform` types to support the
`procedure` field that will be included in v0.8. https://github.com/kwilteam/kwil-db/pull/608

BREAKING CHANGE: The `Database` and `CompiledKuneiform` types changed, as well as many of the
internal checks on action executions. From this commit on, kwil-js should only be used with kwil-db
v0.8+