fix(kwil): separate authentication errors from call errors
There was a minor bug where, when a call failed in private mode, the call error was thrown as part
of the authentication error stack, regardless of whether the call failure was related to
authentication. This commit separates the logic so that only authentication errors are thrown in the
authentication stack.